+++ CHANGELOG.md
@@ -12,6 +12,10 @@ There are 6 types of changes :
 
 # Changes
 
+## 2023-03-03
+### [Changed]
+* I activated Alt keys for prompt
+
 ## 2023-02-26
 ### [Changed]
 * I made focus window recognizable